Switched from rq to celery. Split the event task handling to a different file Added file for logging
